chore(architecture): delete old file and rename components folder name
