+++ edgar-app/src/components/app/connectionPages/ConnectionEdgarCard.tsx
