+++ edgar-app/src/components/app/onboardingPages/OnboardingEdgarCard.tsx
