+++ edgar-app/src/components/app/onboardingPages/medical/OnboardingMedicalSmallCard.tsx
