+++ edgar-app/src/components/app/onboardingPages/medical/emptyInfosWarning/OnboardingMedicalDrawer.tsx
