+++ edgar-app/src/components/app/onboardingPages/medical/emptyInfosWarning/OnboardingMedicalModal.tsx
